refactor: application 및 domain layer 보호를 위해 의존성 역전 #15
