+++ src/main/java/com/readup/server/auth/application/SocialAccountService.java
@@ -2,8 +2,8 @@ package com.readup.server.auth.application;
 
 import org.springframework.stereotype.Service;
 
-import com.readup.server.auth.dto.CreateSocialAccountRequest;
 import com.readup.server.auth.domain.SocialAccount;
+import com.readup.server.auth.dto.CreateSocialAccountRequest;
 import com.readup.server.auth.infrastructure.SocialAccountRepository;
 import com.readup.server.user.domain.User;
