+++ src/main/java/com/readup/server/book/domain/repository/BookRepository.java
@@ -1,0 +1,11 @@
+package com.readup.server.book.domain.repository;
+
+import com.readup.server.book.domain.Book;
+
+public interface BookRepository {
+	Book save(Book newBook);
+
+	boolean existsByIsbnOrTitle(String isbn, String title);
+
+	Book getBookById(Long id);
+}
