--- src/main/java/com/readup/server/book/infrastructure/client/bookinfo/BookInfoClientFacade.java
@@ -1,10 +1,0 @@
-package com.readup.server.book.infrastructure.client.bookinfo;
-
-import com.readup.server.book.infrastructure.client.bookinfo.vo.BookInfoVO;
-
-public interface BookInfoClientFacade {
-
-	BookInfoVO getBookInfo(String isbn);
-
-	String getServiceName();
-}
