+++ src/main/java/com/readup/server/book/infrastructure/repository/BookJpaRepository.java
@@ -5,10 +5,11 @@ import java.util.Optional;
 import org.springframework.data.jpa.repository.JpaRepository;
 
 import com.readup.server.book.domain.Book;
+import com.readup.server.book.domain.repository.BookRepository;
 import com.readup.server.common.exception.ErrorCode;
 import com.readup.server.common.exception.RepositoryException;
 
-public interface BookJpaRepository extends JpaRepository<Book, Long> {
+public interface BookJpaRepository extends JpaRepository<Book, Long>, BookRepository {
 
 	boolean existsByIsbnOrTitle(String isbn, String title);
